test(init-shop): split tests into focused files
